python-sdk/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-vault"
version = "0.1.0"
description = "Zero-trust credential manager for AI agents — Python SDK"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "agent-vault contributors"}]

dependencies = [
    "pyrage>=1.2.0",
    "pyyaml>=6.0",
    "gitpython>=3.1",
]

[project.optional-dependencies]
mcp = ["mcp>=1.2.0"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
agent-vault-mcp = "agent_vault.mcp_server:main"

[tool.setuptools.packages.find]
where = ["src"]
